requirements_nibi.txt
absl_py==2.3.1+computecanada
annotated_types==0.7.0+computecanada
array_record==0.8.1
asttokens==3.0.0+computecanada
astunparse==1.6.3+computecanada
attrs==25.4.0+computecanada
beautifulsoup4==4.14.2+computecanada
cachetools==6.2.2
certifi==2025.11.12
cfgv==3.4.0+computecanada
charset_normalizer==3.4.4+computecanada
chex==0.1.85+computecanada
click==8.3.0+computecanada
cloudpickle==3.1.2+computecanada
contourpy==1.3.2
cycler==0.12.1+computecanada
decorator==4.4.2+computecanada
distlib==0.4.0+computecanada
distrax==0.1.5+computecanada
dlimp @ git+https://github.com/kvablack/dlimp@5edaa4691567873d495633f2708982b42edf1972
dm-tree==0.1.9
einops==0.8.1+computecanada
etils==1.13.0+computecanada
exceptiongroup==1.3.0+computecanada
executing==2.2.1+computecanada
Farama_Notifications==0.0.4+computecanada
filelock==3.20.0+computecanada
flatbuffers==25.9.23+computecanada
flax==0.7.5+computecanada
fonttools==4.60.1
fsspec==2025.10.0+computecanada
gast==0.6.0+computecanada
gdown==5.2.0+computecanada
gitdb==4.0.12+computecanada
gitpython==3.1.45+computecanada
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0+computecanada
grpcio==1.76.0
gym==0.26.2+computecanada
gym-notices==0.1.0
gymnasium==0.29.1+computecanada
h5py==3.15.1
hf-xet==1.2.0
huggingface_hub==0.36.0+computecanada
identify==2.6.15
idna==3.11+computecanada
imageio==2.37.2+computecanada
imageio_ffmpeg==0.6.0+computecanada
importlib_resources==6.5.2+computecanada
ipython==8.37.0
jax==0.4.20+computecanada
jaxlib==0.4.20+cuda12.cudnn89.computecanada
#-e git+https://github.com/Sunlighted/V-GPS.git@406043911b108d82ae37d054418a9234f62fa95d#egg=jaxrl_m
jedi==0.19.2+computecanada
keras==2.15.0+computecanada
kiwisolver==1.4.9
libclang==18.1.1
#-e git+https://github.com/simpler-env/ManiSkill2_real2sim.git@ef7a4d4fdf4b69f2c2154db5b15b9ac8dfe10682#egg=mani_skill2_real2sim&subdirectory=../../../../SimplerEnv/ManiSkill2_real2sim
Markdown==3.10
markdown_it_py==4.0.0+computecanada
MarkupSafe==3.0.3
matplotlib==3.7.0
matplotlib_inline==0.2.1+computecanada
mdurl==0.1.2+computecanada
mediapy==1.2.4+computecanada
ml_collections==1.1.0+computecanada
ml_dtypes==0.5.4
moviepy==1.0.3+computecanada
msgpack==1.1.2
narwhals==2.11.0
nest_asyncio==1.6.0+computecanada
networkx==3.4.2+computecanada
nodeenv==1.9.1+computecanada
numpy==1.26.4+computecanada
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
oauthlib==3.3.1+computecanada
#-e git+https://github.com/Sunlighted/octo.git@2e531e5c5a1ad328f7ada0d7cacc4c70166a1668#egg=octo&subdirectory=../../octo
OpenEXR==3.4.3
opt_einsum==3.4.0+computecanada
optax==0.1.5+computecanada
orbax-checkpoint==0.4.4
overrides==7.7.0+computecanada
packaging==25.0+computecanada
parso==0.8.5+computecanada
pexpect==4.9.0+computecanada
pillow==12.0.0
platformdirs==4.5.0+computecanada
plotly==6.4.0+computecanada
pre_commit==3.3.3+computecanada
proglog==0.1.12
promise==2.3+computecanada
prompt_toolkit==3.0.52+computecanada
protobuf==4.21.12
psutil==7.1.3
ptyprocess==0.7.0+computecanada
pure_eval==0.2.3+computecanada
pyasn1==0.6.1+computecanada
pyasn1_modules==0.4.2+computecanada
pydantic==2.12.4+computecanada
pydantic_core==2.41.5
pygments==2.19.2+computecanada
pyparsing==3.2.5+computecanada
PySocks==1.7.1+computecanada
python_dateutil==2.9.0.post0+computecanada
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5+computecanada
requests_oauthlib==2.0.0+computecanada
rich==14.2.0+computecanada
rsa==4.9.1+computecanada
rtree==1.4.1
ruckig==0.12.2
safetensors==0.6.2
sapien==2.2.2
scipy==1.12.0+computecanada
sentencepiece==0.2.0+computecanada
sentry-sdk==2.44.0
#-e git+https://github.com/Sunlighted/SimplerEnv.git@c3ad9e6f81827718fcc2ec920819606d7843ce6b#egg=simpler_env&subdirectory=../../SimplerEnv
six==1.17.0+computecanada
smmap==5.0.2+computecanada
soupsieve==2.8+computecanada
stack_data==0.6.3+computecanada
tabulate==0.9.0+computecanada
tensorboard==2.15.2+computecanada
tensorboard_data_server==0.7.2+computecanada
tensorflow==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
tensorflow-probability==0.23.0
tensorflow_addons==0.23.0+computecanada
tensorflow_datasets==4.9.2+computecanada
tensorflow_estimator==2.15.0+computecanada
tensorflow_hub==0.16.1+computecanada
tensorflow_text==2.15.0+computecanada
tensorstore==0.1.45
termcolor==3.2.0+computecanada
tf_keras==2.15.1
tokenizers==0.22.1+computecanada
toml==0.10.2+computecanada
toolz==1.1.0+computecanada
tqdm==4.67.1+computecanada
traitlets==5.14.3+computecanada
transformers==4.57.1+computecanada
transforms3d==0.4.2
trimesh==4.9.0
typeguard==2.13.3+computecanada
typing_extensions==4.15.0+computecanada
typing_inspection==0.4.2+computecanada
urllib3==2.5.0+computecanada
virtualenv==20.35.4
wandb==0.23.0
wcwidth==0.2.14+computecanada
werkzeug==3.1.3+computecanada
wrapt==1.14.2
zipp==3.23.0+computecanada
